remove findroot.py
