+++ triples/gui/sos_manager.py
@@ -12,7 +12,6 @@ from ..utils.text_process import (
     preprocess_text, poly_get_factor_form, poly_get_standard_form,
     degree_of_zero, coefficient_triangle, coefficient_triangle_latex
 )
-from ..utils import findroot
 from ..core.sum_of_squares import sum_of_squares, DEFAULT_CONFIGS
 from ..core.linsos import root_tangents
 
@@ -188,23 +187,23 @@ class SOS_Manager():
         """
         return poly_get_factor_form(poly, perm, **kwargs)
 
-    @classmethod
-    def findroot(cls, poly, grid = None, verbose = True):
-        """
-        Find the roots / local minima of a polynomial.
-        """
-        if not cls.check_poly(poly):
-            return []
+    # @classmethod
+    # def findroot(cls, poly, grid = None, verbose = True):
+    #     """
+    #     Find the roots / local minima of a polynomial.
+    #     """
+    #     if not cls.check_poly(poly):
+    #         return []
 
-        roots = findroot(
-            poly, 
-            most = 5, 
-            grid = grid, 
-            with_tangents = root_tangents
-        )
-        if verbose:
-            print(roots)
-        return roots
+    #     roots = findroot(
+    #         poly, 
+    #         most = 5, 
+    #         grid = grid, 
+    #         with_tangents = root_tangents
+    #     )
+    #     if verbose:
+    #         print(roots)
+    #     return roots
 
     @classmethod
     def sum_of_squares(cls,
